fix(proto): split teams grpc defs
